Replace the Button with a SendToken component
