+++ client/src/components/File/FileList.js
@@ -1,10 +1,9 @@
-import React, { useContext, useState } from 'react';
+import React, { useContext } from 'react';
 import { Query } from 'react-apollo';
 import { gql } from 'apollo-boost';
+import SendToken from './SendToken';
 import UpdateFile from './UpdateFile';
 import DeleteFile from './DeleteFile';
-import { FILE_TOKEN_ENDPOINT } from '../../config';
-import { sendEmail } from '../../helpers';
 import ShowError from '../Common/ShowError';
 import Loading from '../Common/Loading';
 import withStyles from '@material-ui/core/styles/withStyles';
@@ -18,13 +17,10 @@ import ExpansionPanelDetails from '@material-ui/core/ExpansionPanelDetails';
 import ExpansionPanelSummary from '@material-ui/core/ExpansionPanelSummary';
 import ExpansionPanelActions from '@material-ui/core/ExpansionPanelActions';
 import ExpandMoreIcon from '@material-ui/icons/ExpandMore';
-import Button from '@material-ui/core/Button';
 
 import { UserContext } from '../../App';
 
 const FileList = ({ dlfiles, classes }) => {
-  const [tokenSent, setTokenSent] = useState(false);
-
   const currentUser = useContext(UserContext);
   //console.log(currentUser.id);
   let userFiles = dlfiles.filter(
@@ -34,21 +30,6 @@ const FileList = ({ dlfiles, classes }) => {
   // const id = match.params.id;
   //console.log(id);
   console.log(window);
-
-  const handleSendToken = ({ fileName, fileDescription, fileToken }) => {
-    const payload = {
-      userName: currentUser.username,
-      email: currentUser.email,
-      fileName,
-      fileDescription,
-      fileToken
-    };
-    sendEmail(FILE_TOKEN_ENDPOINT, payload)
-      .then(() => {
-        setTokenSent(true);
-      })
-      .catch(error => console.log(error));
-  };
 
   // return (
   //   <Query query={PROFILE_QYERY} variables={{ id }}>
@@ -75,21 +56,13 @@ const FileList = ({ dlfiles, classes }) => {
             <Typography variant='body1'>{dlfile.description}</Typography>
           </ExpansionPanelDetails>
           <ExpansionPanelActions>
-            <Button
-              variant={tokenSent ? 'contained' : 'outlined'}
-              onClick={() =>
-                handleSendToken({
-                  fileName: dlfile.name,
-                  fileDescription: dlfile.description,
-                  fileToken: dlfile.fileToken
-                })
-              }
-              disabled={tokenSent}
-            >
-              {tokenSent ? 'Token Sent' : 'Send Token'}
-            </Button>
-            <UpdateFile dlfile={dlfile} />
-            <DeleteFile dlfile={dlfile} />
+            <UpdateFile key={`update-file-${dlfile.id}`} dlfile={dlfile} />
+            <DeleteFile key={`delete-file-${dlfile.id}`} dlfile={dlfile} />
+            <SendToken
+              key={`send-token-${dlfile.id}`}
+              currentUser={currentUser}
+              dlfile={dlfile}
+            />
           </ExpansionPanelActions>
         </ExpansionPanel>
       ))}
